Add static cast for Linux compilation
Should fix #3

+++ src/mioUtils.cpp
@@ -385,7 +385,7 @@ namespace mioUtils
             
             inData->Get<uint64_t>( inValueIndex, value );
             
-            metaValue = value;
+            metaValue = static_cast<qulonglong>( value );
             break;
          }
